+++ fdm-app/app/routes/signin._index.tsx
@@ -7,8 +7,7 @@ import type {
 } from "react-router"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { SiGithub } from "@icons-pack/react-simple-icons"
-import Autoplay from "embla-carousel-autoplay"
-import { AnimatePresence, motion, MotionConfig, useScroll } from "framer-motion"
+import { AnimatePresence, motion, MotionConfig } from "framer-motion"
 import {
   AlertTriangle,
   ArrowDown,
@@ -18,8 +17,7 @@ import {
   Calculator,
   Camera,
   CheckCircle2,
-  ChevronDown,
-  ChevronsRight,
+  ClipboardList,
   Droplets,
   ExternalLink,
   FileUp,
@@ -32,7 +30,6 @@ import {
   ListChecks,
   Mountain,
   MoveDown,
-  Plus,
   Scale,
   ScanText,
   SearchCheck,
@@ -40,15 +37,16 @@ import {
   Sprout,
   Table2,
   Target,
-  TrendingUp,
   Users,
-  ClipboardList,
 } from "lucide-react"
 import { useEffect, useRef, useState } from "react"
 import { Form, Link, redirect, useSearchParams } from "react-router"
 import { RemixFormProvider, useRemixForm } from "remix-hook-form"
 import { dataWithError, redirectWithSuccess } from "remix-toast"
 import { z } from "zod"
+import { IndicatorCalculationFlow } from "~/components/blocks/signin/indicator-calculation-flow"
+import { IndicatorShowcaseCarousel } from "~/components/blocks/signin/indicator-showcase-carousel"
+import { scrollToTop, StickyHeader } from "~/components/blocks/signin/sticky-header"
 import {
   Accordion,
   AccordionContent,
@@ -64,16 +62,7 @@ import {
   CardHeader,
   CardTitle,
 } from "~/components/ui/card"
-import {
-  Carousel,
-  type CarouselApi,
-  CarouselContent,
-  CarouselItem,
-  CarouselNext,
-  CarouselPrevious,
-} from "~/components/ui/carousel"
 import { FormControl, FormField, FormItem, FormLabel, FormMessage } from "~/components/ui/form"
-import { HoverCard, HoverCardContent, HoverCardTrigger } from "~/components/ui/hover-card"
 import { Input } from "~/components/ui/input"
 import { Separator } from "~/components/ui/separator"
 import { Spinner } from "~/components/ui/spinner"
@@ -178,570 +167,6 @@ export async function loader({ request }: LoaderFunctionArgs) {
   } catch (error) {
     throw handleLoaderError(error)
   }
-}
-
-function scrollToTop() {
-  const reduce = window.matchMedia?.("(prefers-reduced-motion: reduce)")?.matches
-  window.scrollTo({ top: 0, behavior: reduce ? "instant" : "smooth" })
-}
-
-const StickyHeader = () => {
-  const { scrollY } = useScroll()
-  const [isVisible, setIsVisible] = useState(false)
-
-  useEffect(() => {
-    return scrollY.on("change", (latest) => {
-      setIsVisible(latest > 600)
-    })
-  }, [scrollY])
-
-  return (
-    <AnimatePresence>
-      {isVisible && (
-        <motion.header
-          initial={{ y: -100, opacity: 0 }}
-          animate={{ y: 0, opacity: 1 }}
-          exit={{ y: -100, opacity: 0 }}
-          transition={{ duration: 0.3 }}
-          className="bg-background/80 fixed top-0 right-0 left-0 z-50 border-b px-4 py-3 shadow-xs backdrop-blur-md"
-        >
-          <div className="container mx-auto flex max-w-6xl items-center justify-between">
-            <div className="flex items-center gap-2">
-              <div className="flex size-8 items-center justify-center rounded-lg bg-[#122023]">
-                <img className="size-6" src={clientConfig.logomark} alt={clientConfig.name} />
-              </div>
-              <span className="font-semibold">{clientConfig.name}</span>
-            </div>
-            <Button onClick={scrollToTop}>Aanmelden</Button>
-          </div>
-        </motion.header>
-      )}
-    </AnimatePresence>
-  )
-}
-
-type IndicatorShowcaseSlide = {
-  id: string
-  icon: typeof Gauge
-  title: string
-  description: string
-  image: string
-  alt: string
-}
-
-const INDICATOR_SHOWCASE_SLIDES: IndicatorShowcaseSlide[] = [
-  {
-    id: "indicatorenboom",
-    icon: Gauge,
-    title: "Van bedrijfsscore tot indicator",
-    description:
-      "Klap de bedrijfsscore van 78 (Goed) uit tot op indicatorniveau — zoals een fosfaatbeschikbaarheid van 59 binnen de categorie Productie (OBI).",
-    image: "/fdm-screenshot-indicators-1.png",
-    alt: "Uitklapbare BLN-indicatorenboom met scores voor Water, Klimaat, Productie (OBI) en BedrijfsBodemWaterPlan (BBWP)",
-  },
-  {
-    id: "knelpunten",
-    icon: AlertTriangle,
-    title: "Weet waar u moet beginnen",
-    description:
-      "De knelpuntenanalyse rangschikt indicatoren en percelen op impact — hier is grondwateraanvulling (34) het grootste knelpunt, met een link naar de vijf percelen die daar het meest aan bijdragen.",
-    image: "/fdm-screenshot-indicators-2.png",
-    alt: "Knelpuntenanalyse met de zwakste indicatoren en percelen met de grootste negatieve impact",
-  },
-  {
-    id: "maatregelen",
-    icon: ClipboardList,
-    title: "Maatregelen op maat van uw percelen",
-    description:
-      "Selecteer de percelen — hier 2 van 204 — waar een maatregel geldt, stel een start- en einddatum in en volg het effect terug in uw indicatoren.",
-    image: "/fdm-screenshot-measures-1.png",
-    alt: "Dialoog voor het toevoegen van een maatregel aan geselecteerde percelen",
-  },
-]
-
-/**
- * Single-slide, legible showcase for the BLN indicators & maatregelen features.
- *
- * Renders one screenshot at a time at a large, readable size, instead of shrinking
- * all three into a cramped grid where the in-app UI text becomes illegible.
- */
-function IndicatorShowcaseCarousel() {
-  const [api, setApi] = useState<CarouselApi>()
-  const [current, setCurrent] = useState(0)
-  const [autoplayPlugin] = useState(() => {
-    const prefersReducedMotion =
-      typeof window !== "undefined" &&
-      window.matchMedia?.("(prefers-reduced-motion: reduce)")?.matches
-    return prefersReducedMotion
-      ? null
-      : Autoplay({ delay: 4000, stopOnMouseEnter: true, stopOnInteraction: false })
-  })
-
-  useEffect(() => {
-    if (!api) return
-    setCurrent(api.selectedScrollSnap())
-    const onSelect = () => setCurrent(api.selectedScrollSnap())
-    api.on("select", onSelect)
-    return () => {
-      api.off("select", onSelect)
-    }
-  }, [api])
-
-  return (
-    <div className="relative">
-      <Carousel
-        setApi={setApi}
-        opts={{ loop: true }}
-        plugins={autoplayPlugin ? [autoplayPlugin] : undefined}
-        className="mx-auto max-w-4xl"
-      >
-        <CarouselContent>
-          {INDICATOR_SHOWCASE_SLIDES.map((slide) => {
-            const Icon = slide.icon
-            return (
-              <CarouselItem key={slide.id}>
-                <div className="mx-auto max-w-2xl text-center">
-                  <div className="bg-primary/10 text-primary mx-auto mb-4 flex h-12 w-12 items-center justify-center rounded-lg">
-                    <Icon className="h-6 w-6" />
-                  </div>
-                  <h3 className="mb-2 text-xl font-semibold">{slide.title}</h3>
-                  <p className="text-muted-foreground text-sm leading-relaxed">
-                    {slide.description}
-                  </p>
-                </div>
-                <div className="bg-muted/10 mx-auto mt-8 flex min-h-[420px] max-w-3xl items-center justify-center rounded-xl border p-6">
-                  <img
-                    src={slide.image}
-                    alt={slide.alt}
-                    className="bg-background h-auto max-h-[440px] w-auto rounded-md object-contain shadow-sm"
-                    loading="lazy"
-                  />
-                </div>
-              </CarouselItem>
-            )
-          })}
-        </CarouselContent>
-        <CarouselPrevious className="left-2 sm:-left-4 lg:-left-10" />
-        <CarouselNext className="right-2 sm:-right-4 lg:-right-10" />
-      </Carousel>
-      <div className="mt-6 flex items-center justify-center gap-2">
-        {INDICATOR_SHOWCASE_SLIDES.map((slide, index) => (
-          <button
-            key={slide.id}
-            type="button"
-            onClick={() => api?.scrollTo(index)}
-            aria-label={`Ga naar ${slide.title}`}
-            aria-current={current === index}
-            className={cn(
-              "h-1.5 rounded-full transition-all",
-              current === index ? "bg-primary w-6" : "bg-muted-foreground/30 w-1.5",
-            )}
-          />
-        ))}
-      </div>
-    </div>
-  )
-}
-
-type FlowStepTone = "input" | "result" | "final"
-
-/** Extra background copy shown on hover/focus for each node in the calculation flow. */
-const FLOW_STEP_DETAILS = {
-  invoerdata:
-    "Alle data die u invoert of ophaalt voor uw bedrijf (bodemanalyses, teeltregistraties, perceelseigenschappen en locatie) vormt de basis voor elke indicator. U hoeft nooit zelf een afgeleide waarde te berekenen.",
-  status:
-    "Per indicator wordt uit uw invoerdata de actuele bodemtoestand afgeleid.",
-  doel: "De streefwaarde wordt per perceel bepaald op basis van locatie, bodemtype en andere kenmerken, hierdoor kunnen twee percelen met hetzelfde gewas toch een ander doel hebben.",
-  index:
-    "De index (0–100) drukt uit hoe dicht de status bij het doel ligt: hoe kleiner de afwijking, hoe hoger de index.",
-  maatregelen:
-    "Elke maatregel wordt eerst getoetst op toepasbaarheid voor het perceel en de indicator, en heeft daarnaast een vastgestelde effectgrootte.",
-  impact:
-    "De impact is het gecombineerde effect van alle toepasselijke, geselecteerde maatregelen voor deze specifieke indicator.",
-  score:
-    "De index en de impact van uw maatregelen worden samen de indicatorscore die u terugziet op de indicatorenpagina, per perceel en op bedrijfsniveau.",
-} as const
-
-/**
- * A single node card in the indicator-calculation flow diagram. When `detail`
- * is provided, hovering (or focusing, for keyboard users) reveals a richer
- * explanation with a concrete example — the short label stays scannable, the
- * hover-card carries the depth.
- */
-function FlowStep({
-  icon: Icon,
-  title,
-  description,
-  detail,
-  tone = "input",
-  compact = false,
-  className,
-}: {
-  icon: typeof Gauge
-  title: string
-  description: string
-  detail?: string
-  tone?: FlowStepTone
-  compact?: boolean
-  className?: string
-}) {
-  const card = (
-    <div
-      tabIndex={detail ? 0 : undefined}
-      className={cn(
-        "rounded-xl border shadow-sm transition-all",
-        compact ? "p-3.5" : "p-5",
-        tone === "final"
-          ? "bg-primary text-primary-foreground border-primary"
-          : tone === "result"
-            ? "bg-primary/5 border-primary/30"
-            : "bg-card",
-        detail &&
-          "focus-visible:ring-ring cursor-help hover:-translate-y-0.5 hover:shadow-md focus-visible:ring-2 focus-visible:outline-none",
-        className,
-      )}
-    >
-      <div className="flex items-start justify-between gap-2">
-        <div
-          className={cn(
-            "flex items-center justify-center rounded-lg",
-            compact ? "mb-2 h-8 w-8" : "mb-3 h-10 w-10",
-            tone === "final"
-              ? "bg-primary-foreground/10 text-primary-foreground"
-              : "bg-primary/10 text-primary",
-          )}
-        >
-          <Icon className={compact ? "h-4 w-4" : "h-5 w-5"} />
-        </div>
-        {detail && (
-          <Info
-            className={cn(
-              "mt-0.5 h-3.5 w-3.5 shrink-0",
-              tone === "final" ? "text-primary-foreground/50" : "text-muted-foreground/50",
-            )}
-            aria-hidden="true"
-          />
-        )}
-      </div>
-      <h4
-        className={cn(
-          "font-semibold",
-          compact ? "mb-0.5 text-sm" : "mb-1",
-          tone === "final" && !compact && "text-lg",
-        )}
-      >
-        {title}
-      </h4>
-      <p
-        className={cn(
-          compact ? "text-xs leading-snug" : "text-sm leading-relaxed",
-          tone === "final" ? "text-primary-foreground/80" : "text-muted-foreground",
-        )}
-      >
-        {description}
-      </p>
-    </div>
-  )
-
-  if (!detail) return card
-
-  return (
-    <HoverCard openDelay={100} closeDelay={100}>
-      <HoverCardTrigger asChild>{card}</HoverCardTrigger>
-      <HoverCardContent className="w-80">
-        <div className="flex items-center gap-2">
-          <div className="bg-primary/10 text-primary flex h-7 w-7 items-center justify-center rounded-md">
-            <Icon className="h-3.5 w-3.5" />
-          </div>
-          <span className="text-sm font-semibold">{title}</span>
-        </div>
-        <p className="text-muted-foreground mt-2 text-sm leading-relaxed">{detail}</p>
-      </HoverCardContent>
-    </HoverCard>
-  )
-}
-
-/** A small vertical connector between flow steps: a chevron with an optional label. */
-function FlowConnector({ label }: { label: string }) {
-  return (
-    <div className="text-muted-foreground/70 flex flex-col items-center gap-1 py-3">
-      <ChevronDown className="h-4 w-4" aria-hidden="true" />
-      <span className="text-xs">{label}</span>
-    </div>
-  )
-}
-
-/**
- * Vertical flow diagram (mobile/tablet): input data becomes a status, which is
- * measured against a target to produce an index (0–100); selected measures
- * produce an impact; index and impact together form the final indicator score.
- */
-function IndicatorCalculationFlowVertical() {
-  const fadeUp = {
-    initial: { opacity: 0, y: 16 },
-    whileInView: { opacity: 1, y: 0 },
-    viewport: { once: true, margin: "-80px" },
-  }
-
-  return (
-    <div className="mx-auto max-w-3xl">
-      {/* Shared input data → fans out into Status and Doel */}
-      <motion.div
-        {...fadeUp}
-        transition={{ duration: 0.4, ease: [0.16, 1, 0.3, 1] }}
-        className="mx-auto max-w-sm"
-      >
-        <FlowStep
-          icon={LayersIcon}
-          title="Invoerdata"
-          description="Bodemanalyse, teeltgegevens, perceelseigenschappen en locatie: de ruwe data die u al heeft."
-          detail={FLOW_STEP_DETAILS.invoerdata}
-          className="w-full text-center [&>div]:mx-auto"
-        />
-      </motion.div>
-      <FlowConnector label="berekend naar" />
-      <div className="grid gap-6 md:grid-cols-2">
-        <motion.div transition={{ duration: 0.4, delay: 0.1, ease: [0.16, 1, 0.3, 1] }} {...fadeUp}>
-          <FlowStep
-            icon={Gauge}
-            title="Status"
-            description="De actuele toestand van de bodem voor deze indicator."
-            detail={FLOW_STEP_DETAILS.status}
-          />
-        </motion.div>
-
-        <motion.div
-          {...fadeUp}
-          transition={{ duration: 0.4, delay: 0.15, ease: [0.16, 1, 0.3, 1] }}
-        >
-          <FlowStep
-            icon={Target}
-            title="Doel"
-            description="De streefwaarde voor een gezonde bodem, specifiek voor dit perceel."
-            detail={FLOW_STEP_DETAILS.doel}
-          />
-        </motion.div>
-      </div>
-
-      <motion.div
-        {...fadeUp}
-        transition={{ duration: 0.4, delay: 0.2, ease: [0.16, 1, 0.3, 1] }}
-        className="flex flex-col items-center"
-      >
-        <FlowConnector label="status vergeleken met doel" />
-        <FlowStep
-          icon={Scale}
-          title="Index (0–100)"
-          description="Hoe dichter de status bij het doel ligt, hoe hoger de index."
-          detail={FLOW_STEP_DETAILS.index}
-          tone="result"
-          className="w-full max-w-sm text-center [&>div]:mx-auto"
-        />
-      </motion.div>
-
-      {/* Branch 2: selected measures → impact */}
-      <motion.div
-        {...fadeUp}
-        transition={{ duration: 0.4, delay: 0.3, ease: [0.16, 1, 0.3, 1] }}
-        className="mx-auto mt-10 flex max-w-sm flex-col items-center border-t pt-10"
-      >
-        <FlowStep
-          icon={ClipboardList}
-          title="Geselecteerde maatregelen"
-          description="Elke maatregel telt mee naar toepasbaarheid op dit perceel en effectgrootte."
-          detail={FLOW_STEP_DETAILS.maatregelen}
-          className="w-full text-center [&>div]:mx-auto"
-        />
-        <FlowConnector label="bepaalt" />
-        <FlowStep
-          icon={TrendingUp}
-          title="Impact"
-          description="Het gecombineerde effect van uw maatregelen op deze indicator."
-          detail={FLOW_STEP_DETAILS.impact}
-          tone="result"
-          className="w-full text-center [&>div]:mx-auto"
-        />
-      </motion.div>
-
-      {/* Merge: index + impact → final score */}
-      <motion.div
-        {...fadeUp}
-        transition={{ duration: 0.4, delay: 0.4, ease: [0.16, 1, 0.3, 1] }}
-        className="mt-6 flex flex-col items-center"
-      >
-        <div className="text-muted-foreground mb-4 flex items-center gap-3 text-sm font-medium">
-          <span>Index</span>
-          <Plus className="h-4 w-4" aria-hidden="true" />
-          <span>Impact</span>
-        </div>
-        <FlowStep
-          icon={BadgeCheck}
-          title="Indicatorscore"
-          description="De score die u terugziet op de indicatorenpagina, per perceel en op bedrijfsniveau."
-          detail={FLOW_STEP_DETAILS.score}
-          tone="final"
-          className="w-full max-w-sm text-center [&>div]:mx-auto"
-        />
-      </motion.div>
-    </div>
-  )
-}
-
-/**
- * Horizontal flow diagram (desktop): the same calculation, laid out left to
- * right as a genuine flowchart. Arrows pulse in sequence on scroll-into-view
- * so the diagram reads as data flowing through the pipeline, once, on reveal.
- */
-function IndicatorCalculationFlowHorizontal() {
-  const node = (delay: number) => ({
-    initial: { opacity: 0, y: 10 },
-    whileInView: { opacity: 1, y: 0 },
-    viewport: { once: true, margin: "-100px" },
-    transition: { duration: 0.4, delay, ease: [0.16, 1, 0.3, 1] as const },
-  })
-
-  const arrow = (delay: number) => ({
-    initial: { opacity: 0 },
-    whileInView: { opacity: 1, x: [0, 5, 0] },
-    viewport: { once: true, margin: "-100px" },
-    transition: {
-      opacity: { duration: 0.3, delay },
-      x: { duration: 0.7, delay: delay + 0.15, ease: "easeInOut" as const },
-    },
-  })
-
-  return (
-    <div className="hidden lg:block">
-      <div className="grid grid-cols-[1fr_2.5rem_1fr_2.5rem_1fr_2.5rem_1.15fr] items-center gap-x-1 gap-y-6">
-        {/* Shared input data, spanning rows 1–2, fans out into Status and Doel */}
-        <motion.div {...node(0)} className="col-start-1 row-span-2 row-start-1">
-          <FlowStep
-            compact
-            icon={LayersIcon}
-            title="Invoerdata"
-            description="Bodemanalyse, teelt, perceel en locatie."
-            detail={FLOW_STEP_DETAILS.invoerdata}
-          />
-        </motion.div>
-        <motion.div
-          {...arrow(0.1)}
-          className="col-start-2 row-span-2 row-start-1 flex justify-center"
-        >
-          <ChevronsRight className="text-muted-foreground/50 h-5 w-5" aria-hidden="true" />
-        </motion.div>
-
-        {/* Row 1: → Status */}
-        <motion.div {...node(0.15)} className="col-start-3 row-start-1">
-          <FlowStep
-            compact
-            icon={Gauge}
-            title="Status"
-            description="Actuele bodemtoestand."
-            detail={FLOW_STEP_DETAILS.status}
-          />
-        </motion.div>
-
-        {/* Row 2: → Doel */}
-        <motion.div {...node(0.2)} className="col-start-3 row-start-2">
-          <FlowStep
-            compact
-            icon={Target}
-            title="Doel"
-            description="Streefwaarde bodem."
-            detail={FLOW_STEP_DETAILS.doel}
-          />
-        </motion.div>
-
-        {/* Merge rows 1–2 → Index */}
-        <motion.div
-          {...arrow(0.3)}
-          className="col-start-4 row-span-2 row-start-1 flex flex-col items-center justify-center gap-1"
-        >
-          <ChevronsRight className="text-primary/60 h-5 w-5" aria-hidden="true" />
-        </motion.div>
-        <motion.div {...node(0.35)} className="col-start-5 row-span-2 row-start-1">
-          <FlowStep
-            compact
-            icon={Scale}
-            title="Index"
-            description="Score 0–100 t.o.v. doel."
-            detail={FLOW_STEP_DETAILS.index}
-            tone="result"
-          />
-        </motion.div>
-        <motion.div
-          {...arrow(0.45)}
-          className="col-start-6 row-span-2 row-start-1 flex justify-center"
-        >
-          <ArrowRight className="text-muted-foreground/50 h-5 w-5" aria-hidden="true" />
-        </motion.div>
-
-        {/* Row 3: Maatregelen → Impact, bypassing Index straight into the final score */}
-        <motion.div {...node(0.1)} className="col-start-1 row-start-3">
-          <FlowStep
-            compact
-            icon={ClipboardList}
-            title="Maatregelen"
-            description="Toepasbaarheid en effect."
-            detail={FLOW_STEP_DETAILS.maatregelen}
-          />
-        </motion.div>
-        <motion.div {...arrow(0.2)} className="col-start-2 row-start-3 flex justify-center">
-          <ArrowRight className="text-muted-foreground/50 h-5 w-5" aria-hidden="true" />
-        </motion.div>
-        <motion.div {...node(0.25)} className="col-start-3 row-start-3">
-          <FlowStep
-            compact
-            icon={TrendingUp}
-            title="Impact"
-            description="Effect van maatregelen."
-            detail={FLOW_STEP_DETAILS.impact}
-            tone="result"
-          />
-        </motion.div>
-        <motion.div
-          {...arrow(0.35)}
-          className="col-span-3 col-start-4 row-start-3 flex items-center gap-1"
-        >
-          <span
-            aria-hidden="true"
-            className="border-muted-foreground/30 h-px flex-1 border-t border-dashed"
-          />
-          <ArrowRight className="text-muted-foreground/50 h-5 w-5 shrink-0" aria-hidden="true" />
-        </motion.div>
-
-        {/* Final: Index + Impact → Indicatorscore */}
-        <motion.div
-          initial={{ opacity: 0, scale: 0.94 }}
-          whileInView={{ opacity: 1, scale: 1 }}
-          viewport={{ once: true, margin: "-100px" }}
-          transition={{ duration: 0.5, delay: 0.55, ease: [0.16, 1, 0.3, 1] }}
-          className="col-start-7 row-span-3 row-start-1 self-stretch"
-        >
-          <FlowStep
-            icon={BadgeCheck}
-            title="Indicatorscore"
-            description="Zichtbaar per perceel en op bedrijfsniveau."
-            detail={FLOW_STEP_DETAILS.score}
-            tone="final"
-            className="flex h-full flex-col justify-center"
-          />
-        </motion.div>
-      </div>
-    </div>
-  )
-}
-
-/** Renders the vertical flow on mobile/tablet and the horizontal flowchart on desktop. */
-function IndicatorCalculationFlow() {
-  return (
-    <>
-      <div className="lg:hidden">
-        <IndicatorCalculationFlowVertical />
-      </div>
-      <IndicatorCalculationFlowHorizontal />
-    </>
-  )
 }
 
 /**
